+++ lib/processor/condition/resource_test.go
@@ -46,6 +46,9 @@ func (f *fakeMgr) GetCondition(name string) (types.Condition, error) {
 	}
 	return nil, types.ErrConditionNotFound
 }
+func (f *fakeMgr) GetRateLimit(name string) (types.RateLimit, error) {
+	return nil, types.ErrRateLimitNotFound
+}
 func (f *fakeMgr) GetPipe(name string) (<-chan types.Transaction, error) {
 	return nil, types.ErrPipeNotFound
 }