+++ lib/ratelimit/constructor_test.go
@@ -1,0 +1,97 @@
+// Copyright (c) 2018 Ashley Jeffs
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+package ratelimit
+
+import (
+	"encoding/json"
+	"testing"
+
+	"github.com/Jeffail/benthos/lib/log"
+	"github.com/Jeffail/benthos/lib/metrics"
+	yaml "gopkg.in/yaml.v2"
+)
+
+func TestConstructorDescription(t *testing.T) {
+	if len(Descriptions()) == 0 {
+		t.Error("package descriptions were empty")
+	}
+}
+
+func TestConstructorBadType(t *testing.T) {
+	conf := NewConfig()
+	conf.Type = "not_exist"
+
+	if _, err := New(conf, nil, log.Noop(), metrics.Noop()); err == nil {
+		t.Error("Expected error, received nil for invalid type")
+	}
+}
+
+func TestConstructorConfigDefaults(t *testing.T) {
+	conf := []Config{}
+
+	if err := json.Unmarshal([]byte(`[
+		{
+			"type": "local",
+			"local": {
+				"count": 50
+			}
+		}
+	]`), &conf); err != nil {
+		t.Error(err)
+	}
+
+	if exp, act := 1, len(conf); exp != act {
+		t.Errorf("Wrong number of config parts: %v != %v", act, exp)
+		return
+	}
+	if exp, act := "1s", conf[0].Local.Interval; exp != act {
+		t.Errorf("Wrong default interval: %v != %v", act, exp)
+	}
+	if exp, act := 50, conf[0].Local.Count; exp != act {
+		t.Errorf("Wrong overridden count: %v != %v", act, exp)
+	}
+}
+
+func TestConstructorConfigDefaultsYAML(t *testing.T) {
+	conf := []Config{}
+
+	if err := yaml.Unmarshal([]byte(`[
+		{
+			"type": "local",
+			"local": {
+				"count": 50
+			}
+		}
+	]`), &conf); err != nil {
+		t.Error(err)
+	}
+
+	if exp, act := 1, len(conf); exp != act {
+		t.Errorf("Wrong number of config parts: %v != %v", act, exp)
+		return
+	}
+	if exp, act := "1s", conf[0].Local.Interval; exp != act {
+		t.Errorf("Wrong default interval: %v != %v", act, exp)
+	}
+	if exp, act := 50, conf[0].Local.Count; exp != act {
+		t.Errorf("Wrong overridden count: %v != %v", act, exp)
+	}
+}