+++ lib/ratelimit/local.go
@@ -25,19 +25,36 @@ import (
 	"fmt"
 	"sync"
 	"time"
+
+	"github.com/Jeffail/benthos/lib/log"
+	"github.com/Jeffail/benthos/lib/metrics"
+	"github.com/Jeffail/benthos/lib/types"
 )
 
 //------------------------------------------------------------------------------
 
-// Config is a config struct containing rate limit fields.
-type Config struct {
+func init() {
+	Constructors[TypeLocal] = TypeSpec{
+		constructor: NewLocal,
+		description: `
+The local rate limit is a simple X every Y type rate limit that can be shared
+across any number of components within the pipeline.`,
+	}
+}
+
+//------------------------------------------------------------------------------
+
+// LocalConfig is a config struct containing rate limit fields for a local rate
+// limit.
+type LocalConfig struct {
 	Count    int    `json:"count" yaml:"count"`
 	Interval string `json:"interval" yaml:"interval"`
 }
 
-// NewConfig returns a rate limit configuration struct with default values.
-func NewConfig() Config {
-	return Config{
+// NewLocalConfig returns a local rate limit configuration struct with default
+// values.
+func NewLocalConfig() LocalConfig {
+	return LocalConfig{
 		Count:    1000,
 		Interval: "1s",
 	}
@@ -45,10 +62,10 @@ func NewConfig() Config {
 
 //------------------------------------------------------------------------------
 
-// Type is a structure that tracks a rate limit, it can be shared across
+// Local is a structure that tracks a rate limit, it can be shared across
 // parallel processes in order to maintain a maximum rate of a protected
 // resource.
-type Type struct {
+type Local struct {
 	mut         sync.Mutex
 	bucket      int
 	lastRefresh time.Time
@@ -57,31 +74,36 @@ type Type struct {
 	period time.Duration
 }
 
-// New creates a rate limit from a configuration struct. This type is safe to
-// share and call from parallel goroutines.
-func New(conf Config) (*Type, error) {
-	if conf.Count <= 0 {
+// NewLocal creates a local rate limit from a configuration struct. This type is
+// safe to share and call from parallel goroutines.
+func NewLocal(
+	conf Config,
+	mgr types.Manager,
+	logger log.Modular,
+	stats metrics.Type,
+) (types.RateLimit, error) {
+	if conf.Local.Count <= 0 {
 		return nil, errors.New("count must be larger than zero")
 	}
-	period, err := time.ParseDuration(conf.Interval)
+	period, err := time.ParseDuration(conf.Local.Interval)
 	if err != nil {
 		return nil, fmt.Errorf("failed to parse interval: %v", err)
 	}
-	return &Type{
-		bucket:      conf.Count,
+	return &Local{
+		bucket:      conf.Local.Count,
 		lastRefresh: time.Now(),
-		size:        conf.Count,
+		size:        conf.Local.Count,
 		period:      period,
 	}, nil
 }
 
 //------------------------------------------------------------------------------
 
-// Get access to the rate limited resource. Returns a duration or an error if
-// the rate limit check fails. The returned duration is either zero (meaning the
-// resource can be hit) or a reasonable length of time to wait before requesting
+// Access the rate limited resource. Returns a duration or an error if the rate
+// limit check fails. The returned duration is either zero (meaning the resource
+// can be accessed) or a reasonable length of time to wait before requesting
 // again.
-func (r *Type) Get() (time.Duration, error) {
+func (r *Local) Access() (time.Duration, error) {
 	r.mut.Lock()
 	r.bucket--